Remove 'apply' and 'applyIf' monkey patches
I wasn't able to find the call site for this patched version; it not
only calls the fn to the left with the arguments given to the right, but
also copies all non-inherited attributes from the this to the right to
the this to the left. But as far as I can tell, these two objects were
the same in all the function calls!

Also, redefining 'apply' is bad and breaks *all kind of stuff*. So no.