Removes requestedPermissions from auth URL
App-web-auth will not actually read this parameter, instead it will look
at what we store into the redis key. So this must have been documentation.
Except I think this is dangerous - the fact that if the url becomes big
the requestedPermissions go missing is not documented and so rare that
implementers will tend to rely on it not happening. Better to puzzle
people right away and then have them implement something that will work
across versions.