Fixes remaining test
Test broke because we assumed that the amount of users on a server would
always be >0, but sometimes we don't have a number because the server
has zero users.

Also removes the special case for 0 users; this adds complexity with
very little gain. The tests now cover more of the remaining code.